+++ src/lib/ReactDataTable/ReactDataTable.tsx
@@ -70,6 +70,7 @@ const ReactDataTable = <TData,>(props: ReactDataTableProps<TData>) => {
                       key={header.id}
                       onClick={header.column.getToggleSortingHandler()}
                       style={{ ...header.column.columnDef.meta?.headerStyle, ...(header.column.getCanSort() ? { cursor: "pointer" } : {}) }}
+                      className={header.column.columnDef.meta?.headerClassName}
                     >
                       {header.isPlaceholder ? null : flexRender(header.column.columnDef.header, header.getContext())}
 
@@ -182,7 +183,7 @@ const ReactDataTable = <TData,>(props: ReactDataTableProps<TData>) => {
           {table.getRowModel().rows.map((row) => (
             <tr key={row.id} style={rowStyle && rowStyle(row.original)}>
               {row.getVisibleCells().map((cell) => (
-                <td key={cell.id} style={cell.column.columnDef.meta?.cellStyle}>
+                <td key={cell.id} style={cell.column.columnDef.meta?.cellStyle} className={cell.column.columnDef.meta?.cellClassName}>
                   {flexRender(cell.column.columnDef.cell, cell.getContext())}
                 </td>
               ))}
@@ -195,7 +196,13 @@ const ReactDataTable = <TData,>(props: ReactDataTableProps<TData>) => {
               {table.getFooterGroups().map((footerGroup) => (
                 <tr key={footerGroup.id}>
                   {footerGroup.headers.map((header) => (
-                    <th key={header.id}>{header.isPlaceholder ? null : flexRender(header.column.columnDef.footer, header.getContext())}</th>
+                    <th
+                      key={header.id}
+                      style={header.column.columnDef.meta?.footerStyle}
+                      className={header.column.columnDef.meta?.footerClassName}
+                    >
+                      {header.isPlaceholder ? null : flexRender(header.column.columnDef.footer, header.getContext())}
+                    </th>
                   ))}
                 </tr>
               ))}